[CBRD-27140] Review fixes: INT64 saturation bound in the AR-sampling expansion, initialize stats_layout, widen the nominal row-count multiply
- btree_get_stats_with_AR_sampling: the multiply-overflow defence still clamped keys/pkeys to INT_MAX,
  the very bound this change removes; saturate at DB_BIGINT_MAX like the rest of the INT64 chain.
- orc_diskrep_from_record: set DISK_REPR.stats_layout (the struct is only written, and the writer emits
  V1 regardless, but an uninitialized layout must never reach a reader's assert).
- qo_estimate_statistics: (heap_num_pages * DB_PAGESIZE) was an int multiply feeding the now-INT64
  heap_num_objects; compute it in INT64.
- heap_file.h: document that heap_get_num_objects () returns a status while heap_estimate () still
  returns the page count.

+++ src/optimizer/query_graph.c
@@ -5998,7 +5998,7 @@ qo_estimate_statistics (MOP class_mop, CLASS_STATS * statblock)
    */
 
   statblock->heap_num_pages = NOMINAL_HEAP_SIZE (class_mop);
-  statblock->heap_num_objects = (statblock->heap_num_pages * DB_PAGESIZE) / NOMINAL_OBJECT_SIZE (class_mop);
+  statblock->heap_num_objects = ((INT64) statblock->heap_num_pages * DB_PAGESIZE) / NOMINAL_OBJECT_SIZE (class_mop);
 
 }
